refactor: share slack text truncation
(cherry picked from commit f4094ab19e07eece99c2b3fb5db0602bd0562acc)
