+++ src/slack/monitor/events/interactions.ts
@@ -1,6 +1,7 @@
 import type { SlackActionMiddlewareArgs } from "@slack/bolt";
 import type { Block, KnownBlock } from "@slack/web-api";
 import { enqueueSystemEvent } from "../../../infra/system-events.js";
+import { truncateSlackText } from "../../truncate.js";
 import { authorizeSlackSystemEventSender } from "../auth.js";
 import type { SlackMonitorContext } from "../context.js";
 import { escapeSlackMrkdwn } from "../mrkdwn.js";
@@ -53,17 +54,6 @@ type InteractionSummary = InteractionSelectionFields & {
   threadTs?: string;
 };
 
-function truncateInteractionString(
-  value: string,
-  max = SLACK_INTERACTION_STRING_MAX_CHARS,
-): string {
-  const trimmed = value.trim();
-  if (trimmed.length <= max) {
-    return trimmed;
-  }
-  return `${trimmed.slice(0, max - 1)}…`;
-}
-
 function sanitizeSlackInteractionPayloadValue(value: unknown, key?: string): unknown {
   if (value === undefined) {
     return undefined;
@@ -75,7 +65,7 @@ function sanitizeSlackInteractionPayloadValue(value: unknown, key?: string): unk
     return REDACTED_INTERACTION_VALUE;
   }
   if (typeof value === "string") {
-    return truncateInteractionString(value);
+    return truncateSlackText(value, SLACK_INTERACTION_STRING_MAX_CHARS);
   }
   if (Array.isArray(value)) {
     const sanitized = value
